redirected C Algorithms
