+++ algorithms/C/arrays/unique-elements-in-an-array.c
