+++ algorithms/C/linked-lists/Insert-and-delete-beginning.c
