+++ algorithms/C/linked-lists/josephus-problem.c
