+++ algorithms/C/queues/double-ended-queue-using-array.c
