+++ algorithms/C/strings/Permutation-of-String.c
