--- arrays/README.md
@@ -1,8 +1,0 @@
-
-
-# Algorithms related to arrays
-
-### C
-
-1. [Unique Elements in an Array](c-or-cpp/unique-elements-in-an-array.c)
-2. [Even and odd no. ](c-or-cpp/even-and-odd.c)
