--- graphs/README.md
@@ -1,6 +1,0 @@
-# Graphs
-
-### C
-
-1. [Prim's Algorithm](c-or-cpp/Prim's-algorithm.c)
-
