--- linked-lists/README.md
@@ -1,5 +1,0 @@
-# Linked Lists
-
-### C
-
-1. [Josephus Problem Using Circular Linked List](c-or-cpp/josephus-problem.c)
