--- queues/README.md
@@ -1,5 +1,0 @@
-# Queues
-
-### C
-
-1. [Double Ended Queue (using arrays)](c-or-cpp/double-ended-queue-using-array.c)
