--- sorting/README.md
@@ -1,5 +1,0 @@
-# Sorting algorithms
-
-### C
-
-1. [Merge Sort](c-or-cpp/merge-sort.c)
