--- strings/README.md
@@ -1,7 +1,0 @@
-# String Algorithms
-
-### C or C++
-
-1. [Palindrome Check](c-or-cpp/palindrome.c)
-2. [Permutation of String](c-or-cpp/Permutation-of-String.c)
-3. [Count Words](c-or-cpp/count-words.c)
